[MOV] html_builder: create input base option component
In order to make a distinction between text inputs and other inputs, we
create a new class that will be extended if necessary, in order to
allow customization.
